8293122: Revert BsdFileStore; move volume capability check to BsdNativeDispatcher
